CustomUser & Settings
Created the backends.py for the CustomUser model, so we can use email
authentication as well as the username authentication.
Installed the basic settings for the project as well as the TODO's
required.
Created the generic package which contains an untested skeleton for the
formViews and another for the apiViews.